code cleanup, replaced CLAMP and NULL macros in bmpman.
Replace NULL macro with nullptr as it should be in C++11.
Replace usage of CLAMP macro with CAP template.
The macro and the template to the same thing
and thus the template is prefered for its type safety.
Documentation on the macro claims it is faster, but that is untrue.
Code generated will be identical unless types offer very unreasonable implicit conversions.

Declared some variables at point of first use.